feat(Events): use strongly typed event emitters
BREAKING CHANGE: This removes the string based on/off API that was present on most core game classes in favor of strongly typed event emitters that are compatible with Angular bindings.

This makes the API usage simpler, and adds strong types to emitted events making them more reliable to use in the frontend.